requirements.txt
ahk==1.7.5
ahk-binary==2023.9.0
pyyaml==6.0.1